Started moving some of the data access classes around so they better line up with the other module structures
