+++ CoreDataAccessView/src/au/gov/asd/tac/constellation/views/dataaccess/state/DataAccessState.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package au.gov.asd.tac.constellation.views.dataaccess;
+package au.gov.asd.tac.constellation.views.dataaccess.state;
 
 import java.util.ArrayList;
 import java.util.HashMap;
