+++ CoreDataAccessView/test/unit/src/au/gov/asd/tac/constellation/views/dataaccess/io/DataAccessStateIoProviderNGTest.java
@@ -13,10 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package au.gov.asd.tac.constellation.views.dataaccess;
+package au.gov.asd.tac.constellation.views.dataaccess.io;
 
+import au.gov.asd.tac.constellation.views.dataaccess.io.DataAccessStateIoProvider;
 import au.gov.asd.tac.constellation.graph.Attribute;
 import au.gov.asd.tac.constellation.graph.GraphWriteMethods;
+import au.gov.asd.tac.constellation.views.dataaccess.state.DataAccessState;
 import com.fasterxml.jackson.core.JsonFactory;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonNode;
